# Speed up scanning by 7-10%